+++ core/ext.php
@@ -22,6 +22,7 @@ class ext extends \phpbb\extension\base
 		'phpbbgallery.core.notification.image_approved',
 		'phpbbgallery.core.notification.image_not_approved',
 		'phpbbgallery.core.notification.image_moderated',
+		'phpbbgallery.core.notification.image_removed',
 		'phpbbgallery.core.notification.new_comment',
 		'phpbbgallery.core.notification.new_image',
 		'phpbbgallery.core.notification.new_report',