+++ core/language/bg/gallery_notifications.php
@@ -42,8 +42,12 @@ $lang = array_merge($lang, [
 	'NOTIFICATION_TYPE_PHPBBGALLERY_NEW_REPORT' => 'Нов доклад на изображение',
 	'NOTIFICATION_PHPBBGALLERY_NEW_REPORT'      => '<strong>%1$s</strong> докладва изображение',
 	'NOTIFICATION_TYPE_PHPBBGALLERY_IMAGE_MODERATED'       => 'Действия за модериране на галерията',
+	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_APPROVED'   => '<strong>%1$s</strong> одобри изображения в албум <strong>%2$s</strong>',
+	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_REJECTED'   => '<strong>%1$s</strong> отхвърли изображения в албум <strong>%2$s</strong>',
 	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_LOCKED'     => '<strong>%1$s</strong> заключи изображения в албум <strong>%2$s</strong>',
 	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_UNLOCKED'   => '<strong>%1$s</strong> отключи изображения в албум <strong>%2$s</strong>',
 	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_UNAPPROVED' => '<strong>%1$s</strong> върна изображения от албум <strong>%2$s</strong> за одобрение',
 	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_DELETED'    => '<strong>%1$s</strong> премахна изображения от албум <strong>%2$s</strong>',
+	'NOTIFICATION_TYPE_PHPBBGALLERY_IMAGE_REMOVED'         => 'Изображения, премахнати от модераторите',
+	'NOTIFICATION_PHPBBGALLERY_IMAGE_REMOVED'              => 'Изображения в албум <strong>%1$s</strong> бяха премахнати от екипа за модериране',
 ]);