+++ core/language/de/gallery_notifications.php
@@ -42,8 +42,12 @@ $lang = array_merge($lang, [
 	'NOTIFICATION_TYPE_PHPBBGALLERY_NEW_REPORT' => 'Neuer Bild-Report',
 	'NOTIFICATION_PHPBBGALLERY_NEW_REPORT'      => '<strong>%1$s</strong> gemeldetes Bild',
 	'NOTIFICATION_TYPE_PHPBBGALLERY_IMAGE_MODERATED'       => 'Galerie-Moderationsaktionen',
+	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_APPROVED'   => '<strong>%1$s</strong> hat Bilder im Album <strong>%2$s</strong> genehmigt',
+	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_REJECTED'   => '<strong>%1$s</strong> hat Bilder im Album <strong>%2$s</strong> abgelehnt',
 	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_LOCKED'     => '<strong>%1$s</strong> hat Bilder im Album <strong>%2$s</strong> gesperrt',
 	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_UNLOCKED'   => '<strong>%1$s</strong> hat Bilder im Album <strong>%2$s</strong> entsperrt',
 	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_UNAPPROVED' => '<strong>%1$s</strong> hat Bilder im Album <strong>%2$s</strong> zur Genehmigung zurückgestellt',
 	'NOTIFICATION_PHPBBGALLERY_IMAGE_MODERATED_DELETED'    => '<strong>%1$s</strong> hat Bilder aus dem Album <strong>%2$s</strong> entfernt',
+	'NOTIFICATION_TYPE_PHPBBGALLERY_IMAGE_REMOVED'         => 'Von der Moderation entfernte Bilder',
+	'NOTIFICATION_PHPBBGALLERY_IMAGE_REMOVED'              => 'Bilder im Album <strong>%1$s</strong> wurden vom Moderationsteam entfernt',
 ]);